Extract tests type annotations into separate module
