+++ tests/functional_tests/flip_tests/test_definition.py
@@ -2,7 +2,7 @@ from hypothesis import given
 
 from lz.functional import flip
 from tests import strategies
-from tests.utils import FunctionCall
+from tests.hints import FunctionCall
 
 
 @given(strategies.transparent_functions_calls)
