+++ tests/functional_tests/pack_tests/test_pickle.py
@@ -2,8 +2,8 @@ from hypothesis import given
 
 from lz.functional import pack
 from tests import strategies
-from tests.utils import (FunctionCall,
-                         round_trip_pickle)
+from tests.hints import FunctionCall
+from tests.utils import round_trip_pickle
 
 
 @given(strategies.transparent_functions_calls)
